test: reorganize into one file
